Move PersistentList to data-structures project
