+++ subprojects/data-structures/src/main/java/org/gradle/internal/collect/PersistentList.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.execution.plan;
+package org.gradle.internal.collect;
 
 import org.gradle.internal.Cast;
